p2p: migrate tests
