+++ p2p/discovery_test.go
@@ -1,0 +1,73 @@
+// Copyright 2019 ChainSafe Systems (ON) Corp.
+// This file is part of gossamer.
+//
+// The gossamer library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The gossamer library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
+
+package p2p
+
+import (
+	"testing"
+	"time"
+)
+
+// wait time to discover and connect using mdns discovery
+var TestDiscoveryTimeout = 3 * time.Second
+
+// test mdns discovery service (discovers and connects)
+func TestDiscovery(t *testing.T) {
+	configA := &Config{
+		Port:        7001,
+		RandSeed:    1,
+		NoBootstrap: true,
+		NoGossip:    true,
+	}
+
+	nodeA, _, _ := createTestService(t, configA)
+	defer nodeA.Stop()
+
+	nodeA.host.noStatus = true
+
+	configB := &Config{
+		Port:        7002,
+		RandSeed:    2,
+		NoBootstrap: true,
+		NoGossip:    true,
+	}
+
+	nodeB, _, _ := createTestService(t, configB)
+	defer nodeB.Stop()
+
+	nodeB.host.noStatus = true
+
+	time.Sleep(TestDiscoveryTimeout)
+
+	peerCountA := nodeA.host.peerCount()
+	peerCountB := nodeB.host.peerCount()
+
+	if peerCountA != 1 {
+		t.Error(
+			"node A does not have expected peer count",
+			"\nexpected:", 1,
+			"\nreceived:", peerCountA,
+		)
+	}
+
+	if peerCountB != 1 {
+		t.Error(
+			"node B does not have expected peer count",
+			"\nexpected:", 1,
+			"\nreceived:", peerCountB,
+		)
+	}
+}
