+++ p2p/status_test.go
@@ -1,0 +1,80 @@
+// Copyright 2019 ChainSafe Systems (ON) Corp.
+// This file is part of gossamer.
+//
+// The gossamer library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The gossamer library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
+
+package p2p
+
+import (
+	"testing"
+	"time"
+)
+
+// wait time for status messages to be exchanged and handled
+var TestStatusTimeout = time.Second
+
+// test exchange status messages after peer connected
+func TestExchangeStatusMessages(t *testing.T) {
+	configA := &Config{
+		Port:        7001,
+		RandSeed:    1,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeA, _, _ := createTestService(t, configA)
+	defer nodeA.Stop()
+
+	configB := &Config{
+		Port:        7002,
+		RandSeed:    2,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeB, _, _ := createTestService(t, configB)
+	defer nodeB.Stop()
+
+	addrInfoB, err := nodeB.host.addrInfo()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = nodeA.host.connect(*addrInfoB)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	time.Sleep(TestStatusTimeout)
+
+	statusB := nodeA.status.peerConfirmed[nodeB.host.h.ID()]
+	if statusB == false {
+		t.Error(
+			"node A did not receive status message from node B",
+			"\nreceived:", statusB,
+			"\nexpected:", true,
+		)
+	}
+
+	statusA := nodeB.status.peerConfirmed[nodeA.host.h.ID()]
+	if statusA == false {
+		t.Error(
+			"node B did not receive status message from node A",
+			"\nreceived:", statusA,
+			"\nexpected:", true,
+		)
+	}
+}
